Support valve's invalid UTF-8 text-encoding atom (#31)
Valve uses this [1] to specify the window name and they haven't fixed it
in over a year. I hate shipping hacks to support broken software, but it's
better than leaving it broken on principle.

* xcb-icccm.el (xcb:icccm:-atoms): add the UTF-8 atom.
(xcb:unmarshal): allow the UTF-8 atom as an alias for
UTF8_STRING when getting (but not when setting) properties.

[1]: https://github.com/ValveSoftware/steam-for-linux/issues/10854

fixes emacs-exwm/exwm#140